refactor: Clean up main.c and add support for 10-position DIP switch for OS selection
`main.c` has been cleaned up, and the option to use a 10-position DIP switch
(RM3HAF-10R) has been added to allow for selecting more operating systems, if necessary.

Additionally, a constant named `SO_MAX_NUM` has been defined. This constant
must be set to the number of operating systems installed on the machine
to which this TinyUSB device will be connected and used.

These changes enhance the flexibility and scalability of the device,
allowing it to support a wider range of operating system configurations
without requiring firmware modifications for each new OS. The cleanup
of `main.c` also improves code readability and maintainability. The
introduction of `SO_MAX_NUM` provides a clear and configurable way to
manage the number of supported operating systems, making the device
more adaptable to different user environments.